requirements.txt
bcrypt==4.0.1
blessed==1.20.0
cffi==1.15.1
cryptography==41.0.2
decorator==5.1.1
fabric==3.1.0
inquirer==3.1.3
invoke==2.2.0
paramiko==3.2.0
pycparser==2.21
PyNaCl==1.5.0
pypsexec==0.3.0
pyspnego==0.9.1
python-dotenv==1.0.0
python-editor==1.0.4
readchar==4.0.5
six==1.16.0
smbprotocol==1.10.1
tqdm==4.65.0
wcwidth==0.2.6
